renamed /gmap into /location
