+++ src/app/common/index.ts
@@ -1,1 +1,1 @@
-export * from './gmap';
+export * from './location';
